+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Improvements
 * added `logstash-logback-encoder` dependency to enable JSON logging
+* ensured all client, validation and unexpected server errors are logged
 * Add support for tracing (`micrometer-tracing-bridge-otel`, `opentelemetry-exporter-otlp`)
 * Add Spring Security configuration (`spring-boot-starter-security`) requiring HTTP Basic
   authentication for `/actuator/prometheus`